FDO89: Support Windows development using VCExpress (Cleanup warnings)
git-svn-id: http://svn.osgeo.org/fdo/trunk@277 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/OWS/Src/FdoOwsHttpHandler.cpp
@@ -446,7 +446,7 @@ void FdoOwsHttpHandler::Skip( FdoInt64 offset )
     boost::mutex::scoped_lock lock(m_mutex);
     if (offset < 0 && -offset > m_currentRead)
 		throw FdoException::Create(FdoException::NLSGetMessage(FDO_NLSID(FDO_2_BADPARAMETER), "Bad parameter to method."));
-    m_currentRead += offset;
+    m_currentRead += (size_t)offset;
 }
 
 void FdoOwsHttpHandler::Reset()
